Added the FileTracker feature
