+++ src/Main.java
@@ -1,5 +1,9 @@
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.WatchEvent;
+import java.nio.file.WatchEvent.Kind;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -12,9 +16,13 @@ import DataBase.Table.ReadTable;
 import DataBase.Table.Table;
 import DataBase.Table.WriteTable;
 import DataSorter.SortByFilter;
+import DataSorter.FileTracker.*;
+import DataSorter.FileTracker.EntryHandlers.Entry;
+
+import static java.nio.file.StandardWatchEventKinds.*;
 
 public class Main {
-	public static void main(String[] args) throws SQLException {
+	public static void main(String[] args) throws SQLException, IOException {
 		SortByFilter dataSorter = new SortByFilter("C:/");
 //	    LinkedList<String> test = new LinkedList<String>();
 //	    test.add(0);
